requirements.txt
streamlit
google-search-results
phidata
groq